+++ test/fixtures/command_call.rb
@@ -5,3 +5,21 @@ foo.bar barrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrr, bazzzzzzzzzzzzzzzzzzzzzzzzzzzzzz
 -
 foo.bar barrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrrr,
         bazzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz
+%
+expect(foo).to receive(fooooooooooooooooooooooooooooooooooooooooooooooooooooooooo)
+-
+expect(foo).to receive(
+  fooooooooooooooooooooooooooooooooooooooooooooooooooooooooo
+)
+%
+expect(foo).not_to receive(fooooooooooooooooooooooooooooooooooooooooooooooooooooooooo)
+-
+expect(foo).not_to receive(
+  fooooooooooooooooooooooooooooooooooooooooooooooooooooooooo
+)
+%
+expect(foo).to_not receive(fooooooooooooooooooooooooooooooooooooooooooooooooooooooooo)
+-
+expect(foo).to_not receive(
+  fooooooooooooooooooooooooooooooooooooooooooooooooooooooooo
+)